move MetadaFilenames class of constants from ComputedFile model to metadata_file module, move output metadata file naming logic from ComputedFile to metadata_file
