+++ api/scpca_portal/test/management/commands/test_load_data.py
@@ -6,7 +6,7 @@ from zipfile import ZipFile
 
 from django.test import TransactionTestCase
 
-from scpca_portal import common
+from scpca_portal import common, metadata_file
 from scpca_portal.management.commands.load_data import Command
 from scpca_portal.models import ComputedFile, Project, ProjectSummary, Sample
 
@@ -491,7 +491,7 @@ class TestLoadData(TransactionTestCase):
         )
         with ZipFile(project_zip_path) as project_zip:
             sample_metadata = project_zip.read(
-                ComputedFile.MetadataFilenames.SINGLE_CELL_METADATA_FILE_NAME
+                metadata_file.MetadataFilenames.SINGLE_CELL_METADATA_FILE_NAME
             )
             sample_metadata_lines = [
                 sm for sm in sample_metadata.decode("utf-8").split("\r\n") if sm
@@ -638,7 +638,7 @@ class TestLoadData(TransactionTestCase):
         )
         with ZipFile(sample_zip_path) as sample_zip:
             with sample_zip.open(
-                ComputedFile.MetadataFilenames.SINGLE_CELL_METADATA_FILE_NAME, "r"
+                metadata_file.MetadataFilenames.SINGLE_CELL_METADATA_FILE_NAME, "r"
             ) as sample_csv:
                 csv_reader = csv.DictReader(
                     TextIOWrapper(sample_csv, "utf-8"), delimiter=common.TAB
@@ -775,7 +775,7 @@ class TestLoadData(TransactionTestCase):
         )
         with ZipFile(project_zip_path) as project_zip:
             sample_metadata = project_zip.read(
-                ComputedFile.MetadataFilenames.SINGLE_CELL_METADATA_FILE_NAME
+                metadata_file.MetadataFilenames.SINGLE_CELL_METADATA_FILE_NAME
             )
             sample_metadata_lines = [
                 sm for sm in sample_metadata.decode("utf-8").split("\r\n") if sm
@@ -863,7 +863,7 @@ class TestLoadData(TransactionTestCase):
         )
         with ZipFile(sample_zip_path) as sample_zip:
             with sample_zip.open(
-                ComputedFile.MetadataFilenames.SINGLE_CELL_METADATA_FILE_NAME, "r"
+                metadata_file.MetadataFilenames.SINGLE_CELL_METADATA_FILE_NAME, "r"
             ) as sample_csv:
                 csv_reader = csv.DictReader(
                     TextIOWrapper(sample_csv, "utf-8"), delimiter=common.TAB
@@ -894,7 +894,7 @@ class TestLoadData(TransactionTestCase):
         )
         with ZipFile(sample_zip_path) as sample_zip:
             with sample_zip.open(
-                ComputedFile.MetadataFilenames.SINGLE_CELL_METADATA_FILE_NAME, "r"
+                metadata_file.MetadataFilenames.SINGLE_CELL_METADATA_FILE_NAME, "r"
             ) as sample_csv:
                 csv_reader = csv.DictReader(
                     TextIOWrapper(sample_csv, "utf-8"), delimiter=common.TAB
@@ -1009,7 +1009,7 @@ class TestLoadData(TransactionTestCase):
         )
         with ZipFile(project_zip_path) as project_zip:
             spatial_metadata_file = project_zip.read(
-                ComputedFile.MetadataFilenames.SPATIAL_METADATA_FILE_NAME
+                metadata_file.MetadataFilenames.SPATIAL_METADATA_FILE_NAME
             )
             spatial_metadata = [
                 sm for sm in spatial_metadata_file.decode("utf-8").split("\r\n") if sm
@@ -1105,7 +1105,7 @@ class TestLoadData(TransactionTestCase):
         )
         with ZipFile(sample_zip_path) as sample_zip:
             with sample_zip.open(
-                ComputedFile.MetadataFilenames.SPATIAL_METADATA_FILE_NAME, "r"
+                metadata_file.MetadataFilenames.SPATIAL_METADATA_FILE_NAME, "r"
             ) as sample_csv:
                 csv_reader = csv.DictReader(
                     TextIOWrapper(sample_csv, "utf-8"), delimiter=common.TAB
